StackMin
Take in an integer n on the first line and n PUSH and POP commands, each on a new line, then find the minimum integer value in the stack. If the stack is empty print "empty" on a new line.

This solution has O(1) algorithm efficiency. It works off using another stack to track the minimum of your stack, then when requested it simply returns the head of the minimum stack.

As I used 2 stacks here the spacial complexity remains the same (it's O(n) ) as if I had just used one stack and iterated through it and found the minimum that way. However, that algorithm is O(n). Therefor there are no trade offs to be made here. This is the best solution to finding the minimum value in a stack.

Algorithm Efficiency: O(1)
Spacial Complexity: O(n)