QueueViaStack
I implemented a queue in Java using stacks. The insert method to a queue is the same as pushing onto a stack, the only trouble is removing the from of the queue. This is done by popping everything except for the last piece of data, off the stack tracking the data (the main one) to a temporary stack. Then storing the final pop of the main stack in an integer to return later. Before this, the empty main stack needs to retrieve this data back from the temporary one, each pop from the temp is pushed onto the main, thus creating a stack with the bottom removed. Then we return the integer storing the bottom of the main stack.